Move CantBeBlockedByMoreThanOneAttachedEffect to common class with minor changes
