+++ Mage.Sets/src/mage/cards/a/AlphaAuthority.java
@@ -4,20 +4,17 @@ package mage.cards.a;
 import java.util.UUID;
 import mage.abilities.Ability;
 import mage.abilities.common.SimpleStaticAbility;
-import mage.abilities.effects.ContinuousEffectImpl;
 import mage.abilities.effects.Effect;
 import mage.abilities.effects.common.AttachEffect;
+import mage.abilities.effects.common.combat.CantBeBlockedByMoreThanOneAttachedEffect;
 import mage.abilities.effects.common.continuous.GainAbilityAttachedEffect;
 import mage.abilities.keyword.EnchantAbility;
 import mage.abilities.keyword.HexproofAbility;
 import mage.cards.CardImpl;
 import mage.cards.CardSetInfo;
 import mage.constants.*;
-import mage.game.Game;
-import mage.game.permanent.Permanent;
 import mage.target.TargetPermanent;
 import mage.target.common.TargetCreaturePermanent;
-import mage.util.CardUtil;
 
 /**
  * @author LevelX2
@@ -38,7 +35,7 @@ public final class AlphaAuthority extends CardImpl {
 
         // Enchanted creature has hexproof and can't be blocked by more than one creature.
         ability = new SimpleStaticAbility(Zone.BATTLEFIELD, new GainAbilityAttachedEffect(HexproofAbility.getInstance(), AttachmentType.AURA, Duration.WhileOnBattlefield));
-        Effect effect = new CantBeBlockedByMoreThanOneAttachedEffect(AttachmentType.AURA, 1);
+        Effect effect = new CantBeBlockedByMoreThanOneAttachedEffect(AttachmentType.AURA);
         effect.setText("and can't be blocked by more than one creature");
         ability.addEffect(effect);
         this.addAbility(ability);
@@ -51,60 +48,5 @@ public final class AlphaAuthority extends CardImpl {
     @Override
     public AlphaAuthority copy() {
         return new AlphaAuthority(this);
-    }
-}
-
-class CantBeBlockedByMoreThanOneAttachedEffect extends ContinuousEffectImpl {
-
-    protected int amount;
-    protected AttachmentType attachmentType;
-
-    public CantBeBlockedByMoreThanOneAttachedEffect(AttachmentType attachmentType, int amount) {
-        this(attachmentType, amount, Duration.WhileOnBattlefield);
-    }
-
-    public CantBeBlockedByMoreThanOneAttachedEffect(AttachmentType attachmentType, int amount, Duration duration) {
-        super(duration, Outcome.Benefit);
-        this.amount = amount;
-        this.attachmentType = attachmentType;
-        staticText = attachmentType.verb() + " creature can't be blocked by more than " + CardUtil.numberToText(amount) + " creature" + (amount == 1 ? "" : "s");
-    }
-
-    private CantBeBlockedByMoreThanOneAttachedEffect(final CantBeBlockedByMoreThanOneAttachedEffect effect) {
-        super(effect);
-        this.amount = effect.amount;
-        this.attachmentType = effect.attachmentType;
-    }
-
-    @Override
-    public CantBeBlockedByMoreThanOneAttachedEffect copy() {
-        return new CantBeBlockedByMoreThanOneAttachedEffect(this);
-    }
-
-    @Override
-    public boolean apply(Layer layer, SubLayer sublayer, Ability source, Game game) {
-        switch (layer) {
-            case RulesEffects:
-                Permanent attachment = game.getPermanent(source.getSourceId());
-                if (attachment != null && attachment.getAttachedTo() != null) {
-                    Permanent perm = game.getPermanent(attachment.getAttachedTo());
-                    if (perm != null) {
-                        perm.setMaxBlockedBy(amount);
-                        return true;
-                    }
-                }
-                break;
-        }
-        return false;
-    }
-
-    @Override
-    public boolean apply(Game game, Ability source) {
-        return false;
-    }
-
-    @Override
-    public boolean hasLayer(Layer layer) {
-        return layer == Layer.RulesEffects;
     }
 }
