+++ Mage.Sets/src/mage/cards/r/Rope.java
@@ -3,7 +3,7 @@ package mage.cards.r;
 import mage.abilities.Ability;
 import mage.abilities.common.SimpleStaticAbility;
 import mage.abilities.costs.mana.GenericManaCost;
-import mage.abilities.effects.ContinuousEffectImpl;
+import mage.abilities.effects.common.combat.CantBeBlockedByMoreThanOneAttachedEffect;
 import mage.abilities.effects.common.continuous.BoostEquippedEffect;
 import mage.abilities.effects.common.continuous.GainAbilityAttachedEffect;
 import mage.abilities.keyword.EquipAbility;
@@ -12,9 +12,6 @@ import mage.abilities.token.ClueAbility;
 import mage.cards.CardImpl;
 import mage.cards.CardSetInfo;
 import mage.constants.*;
-import mage.game.Game;
-import mage.game.permanent.Permanent;
-import mage.util.CardUtil;
 
 import java.util.UUID;
 
@@ -33,7 +30,7 @@ public final class Rope extends CardImpl {
         Ability boostAbility = new SimpleStaticAbility(new BoostEquippedEffect(1, 2));
         boostAbility.addEffect(new GainAbilityAttachedEffect(ReachAbility.getInstance(), AttachmentType.EQUIPMENT)
                 .setText(", has reach"));
-        boostAbility.addEffect(new CantBeBlockedByMoreThanOneAttachedEffect(AttachmentType.EQUIPMENT, 1)
+        boostAbility.addEffect(new CantBeBlockedByMoreThanOneAttachedEffect(AttachmentType.EQUIPMENT)
                 .setText(", and can't be blocked by more than one creature"));
         this.addAbility(boostAbility);
 
@@ -51,61 +48,5 @@ public final class Rope extends CardImpl {
     @Override
     public Rope copy() {
         return new Rope(this);
-    }
-}
-
-// Copied from Alpha Authority
-class CantBeBlockedByMoreThanOneAttachedEffect extends ContinuousEffectImpl {
-
-    protected int amount;
-    protected AttachmentType attachmentType;
-
-    public CantBeBlockedByMoreThanOneAttachedEffect(AttachmentType attachmentType, int amount) {
-        this(attachmentType, amount, Duration.WhileOnBattlefield);
-    }
-
-    public CantBeBlockedByMoreThanOneAttachedEffect(AttachmentType attachmentType, int amount, Duration duration) {
-        super(duration, Outcome.Benefit);
-        this.amount = amount;
-        this.attachmentType = attachmentType;
-        staticText = attachmentType.verb() + " creature can't be blocked by more than " + CardUtil.numberToText(amount) + " creature" + (amount == 1 ? "" : "s");
-    }
-
-    private CantBeBlockedByMoreThanOneAttachedEffect(final CantBeBlockedByMoreThanOneAttachedEffect effect) {
-        super(effect);
-        this.amount = effect.amount;
-        this.attachmentType = effect.attachmentType;
-    }
-
-    @Override
-    public CantBeBlockedByMoreThanOneAttachedEffect copy() {
-        return new CantBeBlockedByMoreThanOneAttachedEffect(this);
-    }
-
-    @Override
-    public boolean apply(Layer layer, SubLayer sublayer, Ability source, Game game) {
-        switch (layer) {
-            case RulesEffects:
-                Permanent attachment = game.getPermanent(source.getSourceId());
-                if (attachment != null && attachment.getAttachedTo() != null) {
-                    Permanent perm = game.getPermanent(attachment.getAttachedTo());
-                    if (perm != null) {
-                        perm.setMaxBlockedBy(amount);
-                        return true;
-                    }
-                }
-                break;
-        }
-        return false;
-    }
-
-    @Override
-    public boolean apply(Game game, Ability source) {
-        return false;
-    }
-
-    @Override
-    public boolean hasLayer(Layer layer) {
-        return layer == Layer.RulesEffects;
     }
 }
