+++ Mage.Sets/src/mage/cards/v/VorracBattlehorns.java
@@ -5,8 +5,8 @@ import java.util.UUID;
 import mage.abilities.Ability;
 import mage.abilities.common.SimpleStaticAbility;
 import mage.abilities.costs.mana.GenericManaCost;
-import mage.abilities.effects.ContinuousEffectImpl;
 import mage.abilities.effects.Effect;
+import mage.abilities.effects.common.combat.CantBeBlockedByMoreThanOneAttachedEffect;
 import mage.abilities.effects.common.continuous.GainAbilityAttachedEffect;
 import mage.abilities.keyword.EquipAbility;
 import mage.abilities.keyword.TrampleAbility;
@@ -15,14 +15,8 @@ import mage.cards.CardSetInfo;
 import mage.constants.AttachmentType;
 import mage.constants.CardType;
 import mage.constants.SubType;
-import mage.constants.Duration;
-import mage.constants.Layer;
 import mage.constants.Outcome;
-import mage.constants.SubLayer;
 import mage.constants.Zone;
-import mage.game.Game;
-import mage.game.permanent.Permanent;
-import mage.util.CardUtil;
 
 /**
  *
@@ -36,7 +30,7 @@ public final class VorracBattlehorns extends CardImpl {
 
         // Equipped creature has trample and can't be blocked by more than one creature.
         Ability ability = new SimpleStaticAbility(Zone.BATTLEFIELD, new GainAbilityAttachedEffect(TrampleAbility.getInstance(), AttachmentType.EQUIPMENT));
-        Effect effect = new CantBeBlockedByMoreThanOneAttachedEffect(AttachmentType.EQUIPMENT, 1);
+        Effect effect = new CantBeBlockedByMoreThanOneAttachedEffect(AttachmentType.EQUIPMENT);
         effect.setText("and can't be blocked by more than one creature");
         ability.addEffect(effect);
         this.addAbility(ability);
@@ -52,60 +46,5 @@ public final class VorracBattlehorns extends CardImpl {
     @Override
     public VorracBattlehorns copy() {
         return new VorracBattlehorns(this);
-    }
-}
-
-class CantBeBlockedByMoreThanOneAttachedEffect extends ContinuousEffectImpl {
-
-    protected int amount;
-    protected AttachmentType attachmentType;
-
-    public CantBeBlockedByMoreThanOneAttachedEffect(AttachmentType attachmentType, int amount) {
-        this(attachmentType, amount, Duration.WhileOnBattlefield);
-    }
-
-    public CantBeBlockedByMoreThanOneAttachedEffect(AttachmentType attachmentType, int amount, Duration duration) {
-        super(duration, Outcome.Benefit);
-        this.amount = amount;
-        this.attachmentType = attachmentType;
-        staticText = attachmentType.verb() + " creature can't be blocked by more than " + CardUtil.numberToText(amount) + " creature" + (amount==1 ?"":"s");
-    }
-
-    private CantBeBlockedByMoreThanOneAttachedEffect(final CantBeBlockedByMoreThanOneAttachedEffect effect) {
-        super(effect);
-        this.amount = effect.amount;
-        this.attachmentType = effect.attachmentType;
-    }
-
-    @Override
-    public CantBeBlockedByMoreThanOneAttachedEffect copy() {
-        return new CantBeBlockedByMoreThanOneAttachedEffect(this);
-    }
-
-    @Override
-    public boolean apply(Layer layer, SubLayer sublayer, Ability source, Game game) {
-        switch (layer) {
-            case RulesEffects:
-                Permanent attachment = game.getPermanent(source.getSourceId());
-                if (attachment != null && attachment.getAttachedTo() != null) {
-                    Permanent perm = game.getPermanent(attachment.getAttachedTo());
-                    if (perm != null) {
-                        perm.setMaxBlockedBy(amount);
-                        return true;
-                    }
-                }
-                break;
-        }
-        return false;
-    }
-
-    @Override
-    public boolean apply(Game game, Ability source) {
-        return false;
-    }
-
-    @Override
-    public boolean hasLayer(Layer layer) {
-        return layer == Layer.RulesEffects;
     }
 }
